# swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: SDX LC
  description: "You can find\nout more about Swagger at\n[http://swagger.io](http://swagger.io)\
    \ or on\n[irc.freenode.net, #swagger](http://swagger.io/irc/).\n"
  termsOfService: http://swagger.io/terms/
  contact:
    email: yxin@renci.org
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
- url: https://virtserver.swaggerhub.com/YufengXin/SDX-LC/1.0.0
  description: SwaggerHub API Auto Mocking
- url: https://0.0.0.0:8080/v2
  description: Localhost
...
paths:
  /topology:
    get:
      tags:
      - topology
      summary: get an existing topology
      description: ID of the topology
      operationId: get_topology
      responses:
        "200":
          description: ok
          content:
            text/plain:
              schema:
                type: string
                x-content-type: text/plain
        "400":
          description: Invalid ID supplied
        "404":
          description: Topology not found
      x-openapi-router-controller: swagger_server.controllers.topology_controller
    put:
      tags:
      - topology
      summary: Update an existing topology
      description: "ID of topology that needs to be updated. \\\\ The topology update\
        \ only updates the addition or deletion \\\\ of node, port, link."
      operationId: update_topology
      requestBody:
        $ref: '#/components/requestBodies/topology'
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: Topology not found
        "405":
          description: Validation exception
      security:
      - topology_auth:
        - write:topology
        - read:topology
      x-openapi-router-controller: swagger_server.controllers.topology_controller
    post:
      tags:
      - topology
      summary: Send a new topology to SDX-LC
      operationId: add_topology
      requestBody:
        $ref: '#/components/requestBodies/topology'
      responses:
        "405":
          description: Invalid input
      security:
      - topology_auth:
        - write:topology
        - read:topology
      x-openapi-router-controller: swagger_server.controllers.topology_controller
    delete:
      tags:
      - topology
      summary: Deletes a topology
      operationId: delete_topology
      parameters:
      - name: api_key
        in: header
        required: false
        style: simple
        explode: false
        schema:
          type: string
      - name: topologyId
        in: query
        description: ID of topology to delete
        required: true
        style: form
        explode: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: topology not found
      security:
      - topology_auth:
        - write:topology
        - read:topology
      x-openapi-router-controller: swagger_server.controllers.topology_controller
...
  /topology/{version}:
    get:
      tags:
      - topology
      summary: Find topology by version
      description: Returns a single topology
      operationId: get_topologyby_version
      parameters:
      - name: topologyId
        in: query
        description: ID of topology to return
        required: true
        style: form
        explode: true
        schema:
          type: integer
          format: int64
      - name: version
        in: path
        description: version of topology to return
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/topology'
        "400":
          description: Invalid version supplied
        "404":
          description: Topology not found
      security:
      - api_key: []
      x-openapi-router-controller: swagger_server.controllers.topology_controller
    delete:
      tags:
      - topology
      summary: Deletes a topology version
      operationId: delete_topology_version
      parameters:
      - name: api_key
        in: header
        required: false
        style: simple
        explode: false
        schema:
          type: string
      - name: topologyId
        in: query
        description: ID of topology to return
        required: true
        style: form
        explode: true
        schema:
          type: integer
          format: int64
      - name: version
        in: path
        description: topology version to delete
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: topology not found
      security:
      - topology_auth:
        - write:topology
        - read:topology
      x-openapi-router-controller: swagger_server.controllers.topology_controller
  /topology/{topologyId}/uploadImage:
    post:
      tags:
      - topology
      summary: uploads an topology image
      operationId: upload_file
      parameters:
      - name: topologyId
        in: path
        description: ID of topology to update
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
      - topology_auth:
        - write:topology
        - read:topology
      x-openapi-router-controller: swagger_server.controllers.topology_controller
  /node:
    get:
      tags:
      - node
      summary: get an existing node
      description: ID of the node
      operationId: get_node
      responses:
        "200":
          description: ok
        "400":
          description: Invalid ID supplied
        "404":
          description: Topology not found
      x-openapi-router-controller: swagger_server.controllers.node_controller
    put:
      tags:
      - topology
      summary: Update an existing node
      description: ID of node that needs to be updated.
      operationId: update_node
      requestBody:
        $ref: '#/components/requestBodies/node'
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: Topology not found
        "405":
          description: Validation exception
      security:
      - topology_auth:
        - write:topology
        - read:topology
      x-openapi-router-controller: swagger_server.controllers.topology_controller
    post:
      tags:
      - node
      summary: add a new node to the topology
      operationId: add_node
      requestBody:
        $ref: '#/components/requestBodies/node'
      responses:
        "405":
          description: Invalid input
      security:
      - topology_auth:
        - write:topology
        - read:topology
      x-openapi-router-controller: swagger_server.controllers.node_controller
    delete:
      tags:
      - node
      summary: Deletes a node
      operationId: delete_node
      parameters:
      - name: api_key
        in: header
        required: false
        style: simple
        explode: false
        schema:
          type: string
      - name: nodeId
        in: query
        description: ID of node to delete
        required: true
        style: form
        explode: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: node not found
      security:
      - topology_auth:
        - write:topology
        - read:topology
      x-openapi-router-controller: swagger_server.controllers.node_controller
...
components:
  schemas:
...
    topology:
      required:
      - id
      - links
      - name
      - nodes
      - time_stamp
      - version
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: amLight
        version:
          type: integer
          format: int64
          example: 1
        time_stamp:
          type: string
          format: date-time
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/node'
        links:
          type: array
          items:
            $ref: '#/components/schemas/link'
      example:
        nodes:
        - name: miami.amLight.sdx
          short_name: Miami-Noviflow
          id: id
          ports:
          - node: null
            name: eth1.miami-Boca.amLight.sdx
            label_range:
            - label_range
            - label_range
            short_name: eth1
            id: id
          - node: null
            name: eth1.miami-Boca.amLight.sdx
            label_range:
            - label_range
            - label_range
            short_name: eth1
            id: id
        - name: miami.amLight.sdx
          short_name: Miami-Noviflow
          id: id
          ports:
          - node: null
            name: eth1.miami-Boca.amLight.sdx
            label_range:
            - label_range
            - label_range
            short_name: eth1
            id: id
          - node: null
            name: eth1.miami-Boca.amLight.sdx
            label_range:
            - label_range
            - label_range
            short_name: eth1
            id: id
        time_stamp: 2000-01-23T04:56:07.000+00:00
        name: amLight
        links:
        - name: miami-Boca.amLight.sdx
          short_name: Miami-BocaRaton
          id: id
          ports:
          - null
          - null
        - name: miami-Boca.amLight.sdx
          short_name: Miami-BocaRaton
          id: id
          ports:
          - null
          - null
        id: id
        version: 1
      xml:
        name: topology
    node:
      required:
      - id
      - location
      - name
      - ports
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: miami.amLight.sdx
        short_name:
          type: string
          example: Miami-Noviflow
        ports:
          type: array
          items:
            $ref: '#/components/schemas/port'
      example:
        name: miami.amLight.sdx
        short_name: Miami-Noviflow
        id: id
        ports:
        - node: null
          name: eth1.miami-Boca.amLight.sdx
          label_range:
          - label_range
          - label_range
          short_name: eth1
          id: id
        - node: null
          name: eth1.miami-Boca.amLight.sdx
          label_range:
          - label_range
          - label_range
          short_name: eth1
          id: id
      xml:
        name: node
    link:
      required:
      - id
      - name
      - nodes
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: miami-Boca.amLight.sdx
        short_name:
          type: string
          example: Miami-BocaRaton
        ports:
          type: array
          items:
            $ref: '#/components/schemas/port'
      example:
        name: miami-Boca.amLight.sdx
        short_name: Miami-BocaRaton
        id: id
        ports:
        - null
        - null
      xml:
        name: link
    port:
      required:
      - id
      - label
      - label_range
      - name
      - node
      - status
      - type
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: eth1.miami-Boca.amLight.sdx
        short_name:
          type: string
          example: eth1
        node:
          $ref: '#/components/schemas/node'
        label_range:
          type: array
          items:
            type: string
      example:
        node: null
        name: eth1.miami-Boca.amLight.sdx
        label_range:
        - label_range
        - label_range
        short_name: eth1
        id: id
      xml:
        name: port
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
      example:
        code: 0
        type: type
        message: message
  requestBodies:
    topology:
      description: topology object that needs to be sent to the SDX LC
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/topology'
      required: true
    node:
      description: node object that needs to be sent to the SDX LC
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/node'
      required: true
...
  securitySchemes:
    topology_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: http://0.0.0.0:8080/oauth/dialog
          scopes:
            write:topology: modify topology in your account
            read:topology: read your topology
      x-tokenInfoFunc: swagger_server.controllers.authorization_controller.check_topology_auth
      x-scopeValidateFunc: swagger_server.controllers.authorization_controller.validate_scope_topology_auth
    api_key:
      type: apiKey
      name: api_key
      in: header
      x-apikeyInfoFunc: swagger_server.controllers.authorization_controller.check_api_key